requirements.txt
Appium-Python-Client~=1.0.2
selenium~=3.141.0
pytest~=6.0.2
pytest-parallel
pytest-rerunfailures
pytest-sugar
pytest-json
requests~=2.24.0
invoke~=1.4.1
python-dotenv
pytest-html-reporter~=0.2.2